This commit was generated by cvs2svn to compensate for changes in r4611, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@4612 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/riotcore.h
@@ -43,7 +43,7 @@
  * Local variables
  */
 
-static int riotio[4];		/* I/O register data */
+static BYTE riotio[4];		/* I/O register data */
 
 static BYTE oldpa;              /* the actual output on PA (input = high) */
 static BYTE oldpb;              /* the actual output on PB (input = high) */
